CameraLightSensor: return START_STICKY
Both codes are only relevant when the phone runs out of memory and kills the service before it finishes executing. START_STICKY tells the OS to recreate the service after it has enough memory and call onStartCommand() again with a null intent. START_NOT_STICKY tells the OS to not bother recreating the service again. There is also a third code START_REDELIVER_INTENT that tells the OS to recreate the service and redeliver the same intent to onStartCommand().

https://stackoverflow.com/questions/9093271/start-sticky-and-start-not-sticky
Signed-off-by: roynatech2544 <whiteshell2544@naver.com>