universal7885: CameraLightSensor: Allow user to mute the app's notification
android.app.NotificationChannel
public void setBlockable(boolean blockable)
Allows users to block notifications sent through this channel, if this channel belongs to a package that is signed with the system signature. If the channel does not belong to a package that is signed with the system signature, this method does nothing, since such channels are blockable by default and cannot be set to be unblockable.

Change-Id: Idfa7a672c0be585381239b879ce386e6e99d0619
Signed-off-by: roynatech2544 <whiteshell2544@naver.com>